feat(ElementPanel): display scope properties for children of DocumentFragments
Sort of related to https://github.com/angular/angular.js/pull/6637 --- Currently,
jqLite#inheritedData() will not find data if it needs to cross through document fragment
barriers, such as the Shadow DOM barrier. This patch allows batarang to figure this out.

In addition, it provides some tests for the angularjs properties sidebar of the element
panel, which were missing previously.

Closes #104