fix(loader): create custom loader that is compatible with angular 1.2 and 1.3
A breaking change with the loading of config blocks between angular 1.2 and 1.3 causes
a an error when registering providers if the appropriate version of angular loader is not
used. This custom loader alters the invokeQueue to use config blocks correctly for both
versions of angular. See the issue with the loader and angular docs:
https://github.com/angular/angular-hint/issues/39 And the breaking change for the
angular loader: https://github.com/angular/angular.js/commit/c0b4e2db9cbc8bc3164cedc4646145d3ab72536e