refactor(hintService): move communication through developer tools to backend service
To improve testability, the backend communication with the developer tools is moved to
a service that can then be mocked for testing purposes. Moving functionality to
a service means that an app file is needed to create the module before the controller
and the service and that the app, service, and controller should be contained in the
application html page.