fix(jsonTree): detect deep model changes via events
Sometimes when navigating around the scopeTree the jsonTree didn't
update.
I believe this was caused by the $watch on the model in jsonTree being
passed the objectEquality flag.
This results in the use of the angular.equal function to detect if the
model has changed.
This function ignores any properties that start with a $...
Thus when only the $properties were different across multiple models,
$watch didn't trigger.
The solution:
a) Use a watch with reference equality to detect when the user selects a
different scope to inspect.
b) Listen for model:change events to detect when the model has been
updated from ngHint.