fix(inspectedApp) detect hints properly
Once Batarang had loaded the hints tab would never update. inspectedApp
was unable to track hints sent to it after model hydration - there was
a bug in the code it used to detect when a message was a hint. Updated
background.js, which does the identification process already, to add an
isHint flag to messages.