fix(inject.js): Appropriately process+filter msgs received via postMessage
The Batarang uses postMessage to send data from the webpage context to the devtools panel, but
the current implementation does not appropriately filter messages that originate from the
webpage itself rather than scripts we inject into the webpage context. As a result, the Batarang
attempts to forward every message the webpage sends via postMessage to the Batarang. I have added
a __fromBatarang property to posted messages, and modified the handler to check for the property
before forwarding it to the devtools panel.

There is also a benign typo in the handler code that references the global event object rather
than the local evt argument. Since these alias, it's not a problem in practice, but fixing it
leads to clearer code.

In addition, the Batarang uses chrome.extension.connect / chrome.extension.sendMessage, when these
methods are supposed to only exist on chrome.runtime according to the Chrome API docs. In
practice, they exist on chrome.extension too, but since this is not a documented API call using
the proper interface is likely to be more future-proof.