fix(background.js) buffer de-serialized model info
After devTools was closed and reopened the jsonTree showed raw json.
This was because background.js was storing the model values as a
json string. Since background.js and inspectedApp now both need the
de-serialized representation of the model, moving the parsing code
into background.js and replacing the string model with the
de-serialized representation on the message itself, thus DRYing up the
code.

Fixes issue #246