fix(jsonTree) array expansion
Intermittently arrays could not be expanded in the json tree. This
occurred because the sort routine for model paths only accounted for
parent child dependency between objects and their dot separated
members and not between arrays and their elements. Updating the sort
function to detect left bracket as path separator as well as dot