fix(jsonTree) show icons
The arrows indicating expanded / collapsed status of objects and arrays
in the jsonTree were missing. The expanded css class was added to a dom
node that was immediately discarded while redrawing the jsonTree to
display the child elements. Updated the code to reapply the css on each
rebuild. Also corrected the relative path to the images.