fix(hint): do not crash
Somewhere between v0.14.4 and v4.0.0, `EventEmitter2` started being inconsistent
wrt to the arguments passed to the `onAny()` listener; sometimes the `eventType`
is passed as the first argument, sometimes it isn't. This caused batarang to
crash.

This commit fixes it, by accounting for the fact that `eventType` might or might
not be passed as a first argument.

Fixes #310